+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # lazy-file CHANGELOG
 
+## HEAD
+
+- Add ability to initialize a LazyFile with `BlobPart[]`, just like a normal `File`
+- Add async iterator support to LazyFile
+
 ## v1.0.0 (Aug 21, 2024)
 
 - Initial release